+++ src/packages/Table/Components/Cells/DateCell.tsx
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 import { CellProps } from 'react-table';
-import { CellRenderProps, TableData } from '../../types';
+import { CellRenderProps, TableData } from '../../Types/types';
 
 export const DateCell = <T extends TableData>(props: CellProps<T, CellRenderProps<T>>) => {
     const {
