+++ src/packages/Table/Components/Cells/NumberCell.tsx
@@ -1,7 +1,7 @@
 import { useMemo } from 'react';
 import { CellProps } from 'react-table';
 import styled from 'styled-components';
-import { CellRenderProps, TableData } from '../../types';
+import { CellRenderProps, TableData } from '../../Types/types';
 
 const Container = styled.div`
     text-align: end;
