+++ src/packages/Table/Components/Cells/RelativeDateCell.tsx
@@ -1,7 +1,7 @@
 import { DateTime } from 'luxon';
 import { useMemo } from 'react';
 import { CellProps } from 'react-table';
-import { CellRenderProps, TableData } from '../../types';
+import { CellRenderProps, TableData } from '../../Types/types';
 
 export const RelativeDateCell = <T extends TableData>(
     props: CellProps<T, CellRenderProps<T>>
