+++ src/packages/Table/Components/TableConfigBar/TableConfigBar.tsx
@@ -1,5 +1,5 @@
+import { TableAPI } from '@equinor/Table';
 import styled from 'styled-components';
-import { TableAPI } from '../Table';
 import { ColumnPicker } from './ColumnPicker/ColumnPicker';
 
 interface TableConfigBarProps {
