+++ src/packages/Table/Utils/utils.tsx
@@ -18,7 +18,7 @@ import {
     TableData,
     HeaderType,
     CustomHeaderType,
-} from '../types';
+} from '../Types/types';
 
 const isCustomHeader = <T extends TableData>(arg: HeaderType<T>): arg is CustomHeaderType<T> => {
     return (arg as CustomHeaderType<T>).Custom !== undefined;
@@ -140,21 +140,21 @@ export const findCellFn = <T extends TableData>(
  */
 export const sortFn =
     <T extends TableData>(key: string): SortByFn<T> =>
-    (objA, objB, id, _desc) => {
-        if (objA?.isGrouped || objA?.isGrouped) {
-            return objA.groupByVal === objB.groupByVal
-                ? 0
-                : objA.groupByVal > objB.groupByVal
-                ? 1
-                : -1;
-        }
-        const a = objA.values[id].content[key];
-        const b = objB.values[id].content[key];
-        if (Number(a) && Number(b)) {
-            return a === b ? 0 : parseFloat(a) > parseFloat(b) ? 1 : -1;
-        }
-        return a === b ? 0 : a > b ? 1 : -1;
-    };
+        (objA, objB, id, _desc) => {
+            if (objA?.isGrouped || objA?.isGrouped) {
+                return objA.groupByVal === objB.groupByVal
+                    ? 0
+                    : objA.groupByVal > objB.groupByVal
+                        ? 1
+                        : -1;
+            }
+            const a = objA.values[id].content[key];
+            const b = objB.values[id].content[key];
+            if (Number(a) && Number(b)) {
+                return a === b ? 0 : parseFloat(a) > parseFloat(b) ? 1 : -1;
+            }
+            return a === b ? 0 : a > b ? 1 : -1;
+        };
 
 /**
  * Pass this to `options={{groupByFn}}` in Table if you need to group columns with complex objects.
